requirements.txt
flask==3.0.3
pillow==10.4.0
numpy==2.1.2
opencv-python==4.10.0.84
